requirements.txt
asgiref>=3.2.10
celery>=4.4.4
channels>=3.0.2
channels-redis>=3.0.1
daphne>=3.0.1
Django>=3.1.4
django-appconf>=1.0.4
django-celery-results>=1.2.1
django-cors-headers>=3.5.0
django-debug-toolbar>=2.2
django-filter>=2.2.0
django-redis>=4.12.1
djangorestframework>=3.12.1
djangorestframework-simplejwt>=4.6.0
mysqlclient>=2.0.1
Pillow>=7.2.0
qrcode>=6.1
python-dateutil>=2.8.1
redis>=3.5.3
requests>=2.23.0
sentry-sdk>=0.19.5
six>=1.14.0
Twisted>=20.3.0
urllib3>=1.25.8
whitenoise>=5.0.1
gunicorn>=19.9.0
pyotp>=2.4.1
phonenumbers>=8.12.21
pymemcache>=3.4.2
django-simple-history>=3.0.0
django-taggit>=1.4.0